Remove conditional search extension from View in SearchConfiguration.swift to streamline code and improve maintainability.
